requirements.txt
pandas==1.1.3
scikit-learn==0.23.2
matplotlib==3.3.2
python-dotenv==0.14.0
jupyterlab==1.2.0
streamlit==0.71.0
